news-backend/requirements.txt
apiclient==1.0.4
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-pythonpath==0.7.3
requests==2.32.3
sqlparse==0.5.1
tomli==2.0.2
typing-extensions==4.12.2
urllib3==2.2.3
